Only execute code when on the command line
If we want to test this and keep it in a single file we need to be able
to require the file in test and not have the code execute.

Grouped all the code into a check for the $0 argument  which when run on the
command line will be the filename, when required there is no $0.