Does not blow up with missing attributes
This script is already in use and is failing when creating the 'alerts'.

We have extracted the code that builds the alerts so we can test it in
isolation.

We did not have any understanding of what the Github API actually
returns so felt our way with what we could see.

The main change is to guard against there being no vulnerable repos and
to use dig so that any missing attributes are nil rather than blowing
up with an error.

We can see the code working now as we got the access key and can see the
missing attributes in the data coming back from Github.

Co-authored-by: MariluzHerrera <maria@dxw.com>