Add an option to filter the results by repo name
This change was motivated by a need by a user to filter their
repositories by standard repo naming conventions. They have teams which
add a prefix to their repo names, while shating the same org. They want
to be able to monitor the security alerts for a single team only, and
filtering by repo name is a good solution for that.