Return empty array when no repos have alerts
Recently we have reached the rare case of having no non-GovPress repos with vulnerability alerts.

This joyous occasion allowed us to discover a bug, where the script would crash on a `nil` where it was expecting an array.

Make sure we always return an array, even if it’s empty.