Do not count fixed alerts
Before: when generating the report, the script did not count alerts that
had been dismissed, but did not account for the fact that alerts that
have been fixed are reported differently in the API (with a "fixedAt"
rather than a "dismissedAt" property). This resulted in it incorrectly
reporting alerts that had in fact been fixed.

Now: the script does not report alerts that have been fixed or
dismissed.