Add CLI option to write data to CSV file
This commit adds a new command line switch: --csv FILE which
causes the script to write data to a CSV file rather than
printing to STDOUT.

This is useful when dealing with a larger number of
vulerabilities in an organisation.

The built in Ruby Csv class is used to handle formatting and
escaping CSV data, and quotes are always used around cell data.
However, some strings may not be parsed well by spreadsheets,
for example "= 9.1.2" may still be read as a forumla.