User can include repos by topic on the command line
This commit adds a new CLI argument --include which
allows users to specify a comma-separated list of
repository topics.

This allows users to include and exclude repositories from
the GraphQL query according to their topics. For example:

    ruby security_alert_notifier.rb --org ORG --include topicA --exclude topicB

would include all repositories configured with topicA from
the query, but exclude repositories configured with topicB.